Moving Human class to its own file
