+++ pyminion/bots/bot.py
@@ -4,7 +4,7 @@ from typing import TYPE_CHECKING, Iterator, List, Optional
 from pyminion.core import CardType, Card
 from pyminion.decider import Decider
 from pyminion.exceptions import CardNotFound, EmptyPile
-from pyminion.players import Player
+from pyminion.player import Player
 
 if TYPE_CHECKING:
     from pyminion.game import Game
