+++ tests/test_cards/test_actions/test_cellar.py
@@ -1,6 +1,6 @@
 from pyminion.expansions.base import Cellar, Copper, Estate, cellar, copper, estate
 from pyminion.game import Game
-from pyminion.players import Human
+from pyminion.human import Human
 
 
 def test_cellar_discard_one(human: Human, game: Game, monkeypatch):
