+++ tests/test_cards/test_actions/test_chapel.py
@@ -1,6 +1,6 @@
 from pyminion.expansions.base import Chapel, Copper, Estate, chapel, copper, estate
 from pyminion.game import Game
-from pyminion.players import Human
+from pyminion.human import Human
 
 
 def test_chapel_trash_one(human: Human, game: Game, monkeypatch):
