+++ tests/test_cards/test_actions/test_remodel.py
@@ -1,6 +1,6 @@
 from pyminion.expansions.base import copper, remodel
 from pyminion.game import Game
-from pyminion.players import Human
+from pyminion.human import Human
 
 
 def test_remodel_gain_valid(human: Human, game: Game, monkeypatch):
