+++ tests/test_cards/test_actions/test_workshop.py
@@ -1,6 +1,6 @@
 from pyminion.expansions.base import workshop
 from pyminion.game import Game
-from pyminion.players import Human
+from pyminion.human import Human
 
 
 def test_workshop_gain_valid(human: Human, game: Game, monkeypatch):
