+++ tests/test_core/test_deck_counter.py
@@ -1,7 +1,8 @@
 from pyminion.core import Deck, DeckCounter
 from pyminion.expansions.base import base_set, copper, estate
 from pyminion.game import Game
-from pyminion.players import Human, Player
+from pyminion.human import Human
+from pyminion.player import Player
 from typing import List
